+++ cardano-testnet/src/Cardano/Testnet.hs
@@ -25,6 +25,10 @@ module Cardano.Testnet (
   makeSocketDir,
   makeTmpBaseAbsPath,
 
+  -- * EpochState processsing helper functions
+  maybeExtractGovernanceActionIndex,
+  findCondition,
+
   -- * Processes
   procChairman,
 
@@ -39,6 +43,7 @@ module Cardano.Testnet (
   ) where
 
 import           Testnet.Components.Query
+import           Testnet.EpochStateProcessing
 import           Testnet.Filepath
 import           Testnet.Process.Run (procChairman)
 import           Testnet.Property.Utils